requirements.txt
validator-collection
discord.py[voice]
spotify-uri
async-spotify
websockets
youtube_dl
requests
pymongo
psutil
rich